Removed the option to look for annotation processor options in the compiler arguments config, since they can not be specified there correctly.
Renamed some methods to make their purpose clearer.

removed the isProcNone method and replaced it with
isAnnotationProcessingEnabled. This now looks for the -proc:none in the
compilerArgument element or the proc config element containing none. If
one of those is present in all configured compilers then annotation
processing is disabled.